# internal/api/controllers/educational_offer_controller.go
package controllers

import (
	"lamsam-web3-backend/internal/consts"
	"lamsam-web3-backend/internal/dto/requests"
	"lamsam-web3-backend/internal/dto/responses"
	"lamsam-web3-backend/internal/models"
	"lamsam-web3-backend/internal/services"
	"lamsam-web3-backend/pkg/security"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/mitchellh/mapstructure"
	"go.uber.org/fx"
)

type EducationalOfferControllerParams struct {
	fx.In
	EducationalOfferService services.EducationalOfferService
}

type EducationalOfferController struct {
	service services.EducationalOfferService
}

func NewEducationalOfferController(p EducationalOfferControllerParams) *EducationalOfferController {
	return &EducationalOfferController{
		service: p.EducationalOfferService,
	}
}

// CreateEducationalOffer godoc
// @Summary Create a new Educational Offer entry
// @Description This endpoint creates a new Educational Offer record. Requires authentication.
// @Tags EducationalOffer
// @Accept  json
// @Produce  json
// @Param Authorization header string true "Bearer Token"
// @Param input body requests.EducationalOfferRequest true "Educational Offer Information"
// @Success 201 {object} models.EducationalOffer "Successfully created"
// @Failure 400 {object} responses.ErrorResponse "Bad request"
// @Failure 401 {object} responses.ErrorResponse "Unauthorized"
// @Failure 500 {object} responses.ErrorResponse "Internal server error"
// @Router /educational-offer/create [post]
func (e *EducationalOfferController) CreateEducationalOffer(ctx *gin.Context) {

	validatedInput, _ := ctx.Get("input")

	educationalOfferRequest := validatedInput.(*requests.EducationalOfferRequest)

	claimsValue, _ := ctx.Get("claims")

	claims, _ := claimsValue.(*security.Claims)

	var educationalOffer models.EducationalOffer
	if err := mapstructure.Decode(educationalOfferRequest, &educationalOffer); err != nil {
		ctx.JSON(http.StatusInternalServerError, responses.ErrorResponse{
			Error: consts.ErrorMapConst,
		})
		return
	}

	createdEducationalOffer, err := e.service.CreateEducationalOffer(&educationalOffer, claims.UserID, educationalOfferRequest.SubtopicIDs)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, responses.ErrorResponse{Error: err.Error()})
		return
	}

	ctx.JSON(http.StatusCreated, createdEducationalOffer)
}